RM-9407 Refactor BootstrapServiceConfigurationTest fixture.
